Change of data format, guess datatypes, do not handle ordinals
Complete change of data ingest format - not backwards-compatible.
We now treat all rows as data rows except the header.
Types are now added Altair-style to column names in the header.
We allow variablename:Q (quantitative) or variablename:N (nominal).
Handling of ordered categoricals ("ordinal") has been removed for now.
Dates are not yet handled.
If a type is not specified, it will be guessed by Pandas. Thus it is usually
enough to specify only the types of those columns that are not obvious from
the data.
We also allow variablename:I (identity column) or variablename:M (metadata).
If more than one identity column is provided, only the first will be used.
Metadata and identity fields are not used in dimensionality reduction, and metadata
fields can be used to colour points.
Types can be specified simultaneously (e.g. variablename:QM).