Toy app: add a server mode using redis and single-user without
Add a flag to run as a "server": multi-threaded, on all interfaces and assuming a redis cache.
Wrap app setup in a function so we can return an initialised server object for running the
server with gunicorn.
Non-server mode now uses a SimpleCache and doesn't require a running redis server.

Also add global variables to control simulated delays.